[refactor] 소셜계정 반환타입 분리
