+++ libs/shared-types/src/lib/res-types/socialAccounts.res.ts
@@ -1,0 +1,8 @@
+export type SocialAccount = {
+  provider: string;
+  registDate: Date;
+  serviceId: string;
+  name: string;
+};
+
+export type SocialAccounts = SocialAccount[];
